9516 Marco is compiled without Xinerama support, 1.18.3 bump
MATE marco is build without XFree Xinerama support, so maximized windows are spread
over both screens. The reason seems to be that `configure` is looking for Solarix
Xinerama on `*-*-solaris*` hosts. As we ship only XFree Xinerama, `configure` won't
bother finding it. Removing the check.

Added couple of build deps which I had to install manually.

Bumped marco to 1.18.3.

Tested that maximized windows work as expected when two screens are on.