Introduce "abuild-common" to deliver an account for Automated Builder implementations
Note: the UID and GID 399 were chosen to stay compatible with SUSE
Open Build System (OBS) standard account since workers in a big build
farm tend to share networked resources and benefit from having same IDs.

On IPS side, there potentially may be a collision with existing user
account on the system where this package is installed (although the
probability is low, assuming that build workers are deployed in zones
or other deployments doctored for the purpose), and pkg(5) feature in
https://github.com/OpenIndiana/pkg5/pull/90 addresses it by ensuring
that such a collision, if it ever happens, does not automatically
introduce a problem into a system.