clang-17: Remove 05-no-default-libgcc_s-linking.patch
libgcc_s must always come before libgcc, so that symbols that are in
both are taken from libgcc_s. The only case where libgcc_s must not be
included is when creating a purely static binary that uses no shared
libraries at all. As soon as any shared library is involved, libgcc_s
*must* be used. Not doing so causes symbols to be taken from libgcc,
which leads to crashes. In particular, with this patch, unwinding is
broken and crashes. Hence, remove it.