Added WindowMaker 0.96.0 as an installable window manager
This commit introduces WindowMaker 0.96.0 as a fully installable window manager for OpenIndiana.
Modifications to the Makefile, windowmaker.p5m, and package files were made to support this version and avoid overwriting existing files. Additionally, the FMRI of 'WindowMaker' was obsoleted in lieu of 'wmaker' for easy searching and package installation.