components/x11/xorg-server: explicitly enable libdrm
Depends on #20404

With the updates to Mesa and libdrm in 2024, glxinfo + glxgears
were in a broken state. Explicitly enabling libdrm within the
xorg-server package fixes this.

With the recent change in default GCC_VERSION to gcc-14, at least
one patch (maybe more) will require tweaking due to gcc-14 now
defaulting to -Werror=sizeof-pointer-memaccess. I have chosen to
set GCC_VERSION to gcc-13 for xorg-server as I will be working on
updating xorg-server to the current upstream release in the near
future.

I am not too happy about needing to pass --disable-dependency-tracking
to pass configure, but for the same reason as above, I don't want to
spend any further time on it right now.