transforms/generate-cleanup: don't package gschemas.compiled
A package I'm working on ran glib-compile-schemas; this left a partial
gschemas.compiled in the proto area (omitting every other package's
schemas).  Through a comedy of errors this ended up listed in the
sample manifest, packaged, and installed, where (being newer than all
the schemas sources) the gconf-cache actuator service didn't rebuild it.

So best to drop it from the sample manifest should it appear.